refactor(release-sidebar): extract header/downloads/activity sections (#14352)
## Summary

Repo-quality loop, iteration 4 (no Linear issue — ad-hoc loop; **Linear workspace is at its free-issue limit**, see below). Behavior-preserving split of `ReleaseSidebar.tsx` (**1113 → 667 lines**).

- The three prop-driven presentation sections (`ReleaseEntityHeader`, `ReleaseArtworkDownloadsSetting`, `ReleaseActivitySection`) + their exclusively-used helpers (`renderArtistLine`, `getReleaseTypeLabel`, `getShellReleaseStatus`, `getDspStatus`, `getProviderGlyph`, `getDspAvatarItems`, `formatTrackCount`, `getPreviewAriaLabel`, `defaultToggleArtworkDownloads`), constants (`RELEASE_TYPE_LABELS`, `DEFAULT_DSP_COLOR`, `RELEASE_SIDEBAR_CARD_CLASSNAME`), and the `ReleaseEntityHeaderProps` type move to a new `ReleaseSidebarSections.tsx` (same `organisms/release-sidebar/` dir — eslint-boundaries safe).
- Only the 3 components are exported; every helper/const/type they use moved **with** them (verified 0 uses in the main body), so nothing else widens. Bodies verbatim. The interleaved DSP-rescan helpers (`formatCooldown`, `getPlatformRescanLabel`, `PLATFORM_RESCAN_COOLDOWN_MS`) stay in `ReleaseSidebar.tsx`.

## Verification

- `pnpm biome check` → **clean ✓**
- `pnpm --filter @jovie/web run typecheck` → **pass ✓** (0 errors)
- `vitest tests/components/organisms/release-sidebar/` → **75/76 pass**. The one failure (`ReleaseSidebarLinks.interaction.test.tsx > adds a DSP link with expected payload shape`) is **pre-existing** — it fails identically on clean `main` with this branch's changes stashed (verified). Not caused by this refactor; unrelated to the moved sections. The dedicated `ReleaseSidebarHeader.test.tsx` (the header this PR moves) passes.
- Pre-commit gate (eslint custom rules incl. boundaries, a11y, typecheck) passed.

## Follow-up (could not file in Linear — usage limit)

The pre-existing `adds a DSP link with expected payload shape` failure should be tracked, but the Linear workspace returned `USAGE_LIMIT_EXCEEDED` (free active-issue limit hit). Flagging here + in the loop STATE so it isn't lost. Repro: stale add-link payload assertion or a timing-flaky waitFor.

## Risks

- Low: same-directory presentation-component move, no logic changes, typecheck green, only pre-existing test red.
